ml_training/rl/run_evaluation.py
"""Comprehensive evaluation script for Edge-RL thesis.

Generates all required plots for Section V of the thesis:
1. RL vs Baselines comparison (Bar chart)
2. Ripening trajectories with agent actions (Time series)
3. Domain randomization envelope (Shaded region)
4. Reward component breakdown (Stacked bar)
"""

import argparse
import json
from pathlib import Path

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
import yaml
from stable_baselines3 import DQN

from ml_training.rl.environment import TomatoRipeningEnv

# Set style
plt.style.use("seaborn-v0_8-whitegrid")
sns.set_context("paper", font_scale=1.5)

def evaluate_trajectories(model, config, n_episodes=5, output_dir=Path("outputs")):
    """Record and plot specific ripening trajectories."""
    env = TomatoRipeningEnv(config=config, seed=42)
    
    # Store data for plotting
    data = []
    
    for i in range(n_episodes):
        obs, info = env.reset()
        done = False
        
        while not done:
            action, _ = model.predict(obs, deterministic=True)
            state = env.simulator.get_state()
            
            data.append({
                "episode": i,
                "hour": state["hours_elapsed"],
                "day": state["days_elapsed"],
                "ripeness": state["_true_ripeness"],
                "temperature": state["_true_temperature"],
                "action": int(action),
                "target_day": env.target_day
            })
            
            obs, _, terminated, truncated, _ = env.step(action)
            done = terminated or truncated

    df = pd.DataFrame(data)
    
    # Plot 1: Ripening Trajectories
    fig, ax = plt.subplots(figsize=(10, 6))
    sns.lineplot(data=df, x="day", y="ripeness", hue="episode", palette="tab10", ax=ax, legend=False)
    
    # Add target zone
    ax.axhspan(3.5, 4.5, color="green", alpha=0.1, label="Optimal Harvest Window")
    ax.axhline(5.0, color="red", linestyle="--", label="Overripe")
    
    ax.set_xlabel("Time (days)")
    ax.set_ylabel("Ripeness Stage (0-5)")
    ax.set_title("RL Agent Ripening Control Trajectories")
    ax.set_ylim(0, 5.5)
    
    plt.tight_layout()
    plt.savefig(output_dir / "ripening_trajectories.png", dpi=300)
    plt.close()
    
    return df

def plot_domain_randomization(model, config, n_episodes=50, output_dir=Path("outputs")):
    """Plot the envelope of trajectories under domain randomization."""
    env = TomatoRipeningEnv(config=config)
    
    trajectories = []
    
    for i in range(n_episodes):
        obs, _ = env.reset()
        traj = []
        done = False
        while not done:
            action, _ = model.predict(obs, deterministic=True)
            state = env.simulator.get_state()
            traj.append(state["_true_ripeness"])
            obs, _, terminated, truncated, _ = env.step(action)
            done = terminated or truncated
        trajectories.append(traj)
    
    # Pad to same length for averaging
    max_len = max(len(t) for t in trajectories)
    padded = np.array([t + [np.nan]*(max_len-len(t)) for t in trajectories])
    
    mean_traj = np.nanmean(padded, axis=0)
    std_traj = np.nanstd(padded, axis=0)
    x = np.linspace(0, max_len/24.0, max_len)
    
    fig, ax = plt.subplots(figsize=(10, 6))
    
    # Plot envelope
    ax.fill_between(x, mean_traj - std_traj, mean_traj + std_traj, alpha=0.3, color="blue", label="±1 Std Dev")
    ax.plot(x, mean_traj, color="blue", linewidth=2, label="Mean Trajectory")
    
    ax.set_xlabel("Time (days)")
    ax.set_ylabel("Ripeness Stage")
    ax.set_title("Robustness to Domain Randomization")
    ax.legend()
    
    plt.tight_layout()
    plt.savefig(output_dir / "domain_randomization_envelope.png", dpi=300)
    plt.close()

def main():
    parser = argparse.ArgumentParser()
    parser.add_argument("--model-path", type=str, required=True)
    parser.add_argument("--config", type=str, default="ml_training/config.yaml")
    args = parser.parse_args()
    
    with open(args.config) as f:
        config = yaml.safe_load(f)
        
    # Load model
    model = DQN.load(args.model_path)
    output_dir = Path(args.model_path).parent
    
    print("Generating ripening trajectories...")
    evaluate_trajectories(model, config, output_dir=output_dir)
    
    print("Generating domain randomization plot...")
    plot_domain_randomization(model, config, output_dir=output_dir)
    
    print(f"All plots saved to {output_dir}")

if __name__ == "__main__":
    main()
