requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
onnx>=1.15.0

# RL
stable-baselines3>=2.1.0
gymnasium>=0.29.0

# Data & Viz
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
pandas>=2.1.0

# Utils
pyyaml>=6.0
tqdm>=4.66.0

# ESP-DL quantization (install separately if needed)
# pip install esp-ppq
